Refactor dealing with empty transcripts
