+++ src/transcribe.py
@@ -4,8 +4,13 @@ from pathlib import Path
 from swescribe.__main__ import pipeline
 from tqdm import tqdm
 
+from blank_transcripts import (
+    load_empty_filenames,
+    remove_empty_transcripts,
+    write_empty_filenames,
+)
 from name_to_path import NameToPathMapper
-from settings import corpus_root, empty_srts_file, video_root
+from settings import video_root
 
 mapper = NameToPathMapper()
 
@@ -52,49 +57,6 @@ def group_to_paths(group, force=False):
         elif not srt_path.exists():
             yield video, srt_path
     write_empty_filenames(empty_files)
-
-
-def emtpy_srt():
-    for srt in corpus_root.glob("**/*.srt"):
-        with open(srt, "r", encoding="utf8") as f:
-            text = f.read()
-        if text.strip() == "":
-            yield srt
-
-
-def load_empty_filenames():
-    if not empty_srts_file.exists():
-        return set()
-    with open(empty_srts_file, "r", encoding="utf8") as f:
-        return {line.strip() for line in f.readlines()}
-
-
-def write_empty_filenames(empyt_ids_set: list):
-    with open(empty_srts_file, "w", encoding="utf8") as f:
-        f.write("\n".join(empyt_ids_set))
-
-
-def remove_empty_transcripts():
-    empty_ids = load_empty_filenames() | {
-        empty_srt.name.replace(".srt", "") for empty_srt in emtpy_srt()
-    }
-
-    write_empty_filenames(sorted(empty_ids))
-
-    for empty_id in empty_ids:
-        files = list(corpus_root.glob(f"**/{empty_id}.srt"))
-        assert len(files) in {0, 1}
-
-        if not files:
-            continue
-
-        for file in files:
-            file.unlink()
-        # delete emtpy dir
-        if len(list(file.parent.iterdir())) == 0:
-            file.parent.rmdir()
-            if len(list(file.parents[1].iterdir())) == 0:
-                file.parents[1].rmdir()
 
 
 if __name__ == "__main__":
